docs: adds comments and reorganizes directories for clarity
